+++ radix-engine/src/model/key_value_store.rs
